ran code cleanup on solution

+++ CSI.BatchTracker/Domain/NativeModels/AverageBatch.cs
@@ -8,7 +8,7 @@
 
         public float AverageUsage
         {
-            get { return (float)QuantityUsed / (float)ProductionDays; }
+            get { return QuantityUsed / (float)ProductionDays; }
         }
 
         public string DisplayUsage
